cast int to unsigned int to fix warining in debug build

+++ protoduneana/singlephase/driftvel/driftvel_module.cc
@@ -315,7 +315,7 @@ namespace protoana{
 	cout<<"vhit vmetadata"<<endl;
 	for (size_t ii = 0; ii<vhit.size(); ++ii){ //loop over all meta data hit
 	  bool fBadhit = false;
-	  if (vmeta[ii]->Index() == std::numeric_limits<int>::max()){
+	  if (vmeta[ii]->Index() == static_cast<unsigned int>(std::numeric_limits<int>::max())){
 	    fBadhit = true;
 	    cout<<"fBadHit"<<fBadhit<<endl;
 	    continue;
